Moves GameState out of logic and removes the folders.
