+++ src/state/GameState.java
@@ -1,4 +1,4 @@
-package logic.game;
+package state;
 
 import factories.MinesFactory;
 import models.Difficulty;
